Move ConvExact to mod traits
